Arreglo de errores y agregaciones de Simulaciones de credito
Se arreglaron errores minimos que eran principalmente por un bloque de codigo que provoco que todo colapse.

Se añadieron metodos y urls para el uso de las simulaciones.

+++ src/main/java/cl/programatufuturo/compararCredito/Comparacion/de/Credito/controllers/CreditoController.java
@@ -25,7 +25,7 @@ public class CreditoController {
 	
 	//Testear
 	@DeleteMapping("deleteCredit")
-	public int deletecredit(@RequestBody int id) {
+	public long deletecredit(@RequestBody long id) {
 		return id;
 	}
